feat: use account-api + split up rules into seperate files
