+++ package.json
@@ -46,7 +46,8 @@
   "resolutions": {
     "elliptic@6.5.4": "^6.5.7",
     "fast-xml-parser@^4.3.4": "^4.4.1",
-    "ws@7.4.6": "^7.5.10"
+    "ws@7.4.6": "^7.5.10",
+    "@metamask/account-api@^0.0.0": "npm:@metamask-previews/account-api@0.0.0-dff9202"
   },
   "devDependencies": {
     "@babel/core": "^7.23.5",
