+++ packages/account-tree-controller/src/AccountTreeController.ts
@@ -1,3 +1,5 @@
+import type { AccountGroupId, AccountWalletId } from '@metamask/account-api';
+import { toDefaultAccountGroupId } from '@metamask/account-api';
 import type {
   AccountId,
   AccountsControllerAccountAddedEvent,
@@ -12,39 +14,18 @@ import {
   BaseController,
 } from '@metamask/base-controller';
 import type { KeyringControllerGetStateAction } from '@metamask/keyring-controller';
-import { KeyringTypes } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 import type { GetSnap as SnapControllerGetSnap } from '@metamask/snaps-controllers';
-import type { SnapId } from '@metamask/snaps-sdk';
-import { stripSnapPrefix } from '@metamask/snaps-utils';
 
-import { getAccountWalletNameFromKeyringType } from './names';
+import type { Rule } from './rules';
+import { EntropySourceRule, SnapIdRule, KeyringTypeRule } from './rules';
 
 const controllerName = 'AccountTreeController';
-
-export enum AccountWalletCategory {
-  Entropy = 'entropy',
-  Keyring = 'keyring',
-  Snap = 'snap',
-}
-
-type AccountTreeRuleMatch = {
-  category: AccountWalletCategory;
-  id: AccountWalletId;
-  name: string;
-};
-
-type AccountTreeRuleFunction = (
-  account: InternalAccount,
-) => AccountTreeRuleMatch | undefined;
 
 type AccountReverseMapping = {
   walletId: AccountWalletId;
   groupId: AccountGroupId;
 };
-
-export type AccountWalletId = `${AccountWalletCategory}:${string}`;
-export type AccountGroupId = `${AccountWalletId}:${string}`;
 
 // Do not export this one, we just use it to have a common type interface between group and wallet metadata.
 type Metadata = {
@@ -55,18 +36,18 @@ export type AccountWalletMetadata = Metadata;
 
 export type AccountGroupMetadata = Metadata;
 
-export type AccountGroup = {
+export type AccountGroupObject = {
   id: AccountGroupId;
   // Blockchain Accounts:
   accounts: AccountId[];
   metadata: AccountGroupMetadata;
 };
 
-export type AccountWallet = {
+export type AccountWalletObject = {
   id: AccountWalletId;
   // Account groups OR Multichain accounts (once available).
   groups: {
-    [groupId: AccountGroupId]: AccountGroup;
+    [groupId: AccountGroupId]: AccountGroupObject;
   };
   metadata: AccountWalletMetadata;
 };
@@ -75,7 +56,7 @@ export type AccountTreeControllerState = {
   accountTree: {
     wallets: {
       // Wallets:
-      [walletId: AccountWalletId]: AccountWallet;
+      [walletId: AccountWalletId]: AccountWalletObject;
     };
   };
 };
@@ -132,49 +113,7 @@ export function getDefaultAccountTreeControllerState(): AccountTreeControllerSta
   };
 }
 
-// TODO: For now we use this for the 2nd-level of the tree until we implements proper multichain accounts.
-export const DEFAULT_ACCOUNT_GROUP_UNIQUE_ID: string = 'default'; // This might need to be re-evaluated based on new structure
 export const DEFAULT_ACCOUNT_GROUP_NAME: string = 'Default';
-
-/**
- * Convert a unique ID to a wallet ID for a given category.
- *
- * @param category - A wallet category.
- * @param id - A unique ID.
- * @returns A wallet ID.
- */
-export function toAccountWalletId(
-  category: AccountWalletCategory,
-  id: string,
-): AccountWalletId {
-  return `${category}:${id}`;
-}
-
-/**
- * Convert a wallet ID and a unique ID, to a group ID.
- *
- * @param walletId - A wallet ID.
- * @param id - A unique ID.
- * @returns A group ID.
- */
-export function toAccountGroupId(
-  walletId: AccountWalletId,
-  id: string,
-): AccountGroupId {
-  return `${walletId}:${id}`;
-}
-
-/**
- * Convert a wallet ID to the default group ID.
- *
- * @param walletId - A wallet ID.
- * @returns The default group ID.
- */
-export function toDefaultAccountGroupId(
-  walletId: AccountWalletId,
-): AccountGroupId {
-  return toAccountGroupId(walletId, DEFAULT_ACCOUNT_GROUP_UNIQUE_ID);
-}
 
 export class AccountTreeController extends BaseController<
   typeof controllerName,
@@ -183,7 +122,7 @@ export class AccountTreeController extends BaseController<
 > {
   readonly #reverse: Map<AccountId, AccountReverseMapping>;
 
-  readonly #rules: AccountTreeRuleFunction[];
+  readonly #rules: Rule[];
 
   /**
    * Constructor for AccountTreeController.
@@ -215,11 +154,11 @@ export class AccountTreeController extends BaseController<
     // Rules to apply to construct the wallets tree.
     this.#rules = [
       // 1. We group by entropy-source
-      (account: InternalAccount) => this.#matchGroupByEntropySource(account),
+      new EntropySourceRule(this.messagingSystem),
       // 2. We group by Snap ID
-      (account: InternalAccount) => this.#matchGroupBySnapId(account),
+      new SnapIdRule(this.messagingSystem),
       // 3. We group by wallet type (this rule cannot fail and will group all non-matching accounts)
-      (account: InternalAccount) => this.#matchGroupByKeyringType(account),
+      new KeyringTypeRule(this.messagingSystem),
     ];
 
     this.messagingSystem.subscribe(
@@ -273,125 +212,12 @@ export class AccountTreeController extends BaseController<
     }
   }
 
-  #hasKeyringType(account: InternalAccount, type: KeyringTypes): boolean {
-    return account.metadata.keyring.type === (type as string);
-  }
-
-  #matchGroupByEntropySource(
-    account: InternalAccount,
-  ): AccountTreeRuleMatch | undefined {
-    let entropySource: string | undefined;
-
-    if (this.#hasKeyringType(account, KeyringTypes.hd)) {
-      // TODO: Maybe use superstruct to validate the structure of HD account since they are not strongly-typed for now?
-      if (!account.options.entropySource) {
-        console.warn(
-          "! Found an HD account with no entropy source: account won't be associated to its wallet",
-        );
-        return undefined;
-      }
-
-      entropySource = account.options.entropySource as string;
-    }
-
-    // TODO: For now, we're not checking if the Snap is a preinstalled one, and we probably should...
-    if (
-      this.#hasKeyringType(account, KeyringTypes.snap) &&
-      account.metadata.snap?.enabled
-    ) {
-      // Not all Snaps have an entropy-source and options are not typed yet, so we have to check manually here.
-      if (account.options.entropySource) {
-        // We blindly trust the `entropySource` for now, but it could be wrong since it comes from a Snap.
-        entropySource = account.options.entropySource as string;
-      }
-    }
-
-    if (!entropySource) {
-      return undefined;
-    }
-
-    // We check if we can get the name for that entropy source, if not this means this entropy does not match
-    // any HD keyrings, thus, is invalid (this account will be grouped by another rule).
-    const entropySourceName = this.#getEntropySourceName(entropySource);
-    if (!entropySourceName) {
-      console.warn(
-        '! Tried to name a wallet using an unknown entropy, this should not be possible.',
-      );
-      return undefined;
-    }
-
-    return {
-      category: AccountWalletCategory.Entropy,
-      id: toAccountWalletId(AccountWalletCategory.Entropy, entropySource),
-      name: entropySourceName,
-    };
-  }
-
-  #matchGroupBySnapId(
-    account: InternalAccount,
-  ): AccountTreeRuleMatch | undefined {
-    if (
-      this.#hasKeyringType(account, KeyringTypes.snap) &&
-      account.metadata.snap &&
-      account.metadata.snap.enabled
-    ) {
-      const { id } = account.metadata.snap;
-
-      return {
-        category: AccountWalletCategory.Snap,
-        id: toAccountWalletId(AccountWalletCategory.Snap, id),
-        name: this.#getSnapName(id as SnapId),
-      };
-    }
-
-    return undefined;
-  }
-
-  #matchGroupByKeyringType(
-    account: InternalAccount,
-  ): AccountTreeRuleMatch | undefined {
-    const { type } = account.metadata.keyring;
-
-    return {
-      category: AccountWalletCategory.Keyring,
-      id: toAccountWalletId(AccountWalletCategory.Keyring, type),
-      name: getAccountWalletNameFromKeyringType(type as KeyringTypes),
-    };
-  }
-
-  #getSnapName(snapId: SnapId): string {
-    const snap = this.messagingSystem.call('SnapController:get', snapId);
-    const snapName = snap
-      ? // TODO: Handle localization here, but that's a "client thing", so we don't have a `core` controller
-        // to refer to.
-        snap.manifest.proposedName
-      : stripSnapPrefix(snapId);
-
-    return snapName;
-  }
-
-  #getEntropySourceName(entropySource: string): string | undefined {
-    const { keyrings } = this.messagingSystem.call(
-      'KeyringController:getState',
-    );
-
-    const index = keyrings
-      .filter((keyring) => keyring.type === (KeyringTypes.hd as string))
-      .findIndex((keyring) => keyring.metadata.id === entropySource);
-
-    if (index === -1) {
-      return undefined;
-    }
-
-    return `Wallet ${index + 1}`; // Use human indexing.
-  }
-
   #insert(
-    wallets: { [walletId: AccountWalletId]: AccountWallet },
+    wallets: { [walletId: AccountWalletId]: AccountWalletObject },
     account: InternalAccount,
   ) {
     for (const rule of this.#rules) {
-      const match = rule(account);
+      const match = rule.match(account);
 
       if (!match) {
         // No match for that rule, we go to the next one.
